agent-service/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
anthropic==0.40.0
aiosqlite==0.20.0
python-dotenv==1.0.1
keyring==25.5.0
apscheduler==3.10.4
supabase==2.13.0

# Dev / testing
httpx==0.27.2
pytest==8.3.4
pytest-asyncio==0.24.0
ruff==0.8.3

# Gmail integration (optional)
google-auth==2.36.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.154.0
